refactor: HI GITBLAMER READ THIS COMMIT MESSAGE; break up util/http.
* hi gitblamer, here is the history you are looking for:
  https://github.com/nafundi/jubilant-garbanzo/blame/d972f04fa024aacc242da8dfd0278ad8c5a97b6e/lib/util/http.js
  * this project will probably get renamed and moved so you may have to
    replace nafundi and jubilant-garbanzo with whatever their current
    equivalent is. also how is the future is it nice?
* in general, util/http got split into three places:
  * lib/http/endpoint now contains all the endpoint wrapper functions.
    * it also contains sendError which may be regarded as a mistake.
      might move it again very soon.
  * lib/http/middleware now contains all the middleware parsing funcs.
  * lib/util/promise absorbed all the getOrX helpers which were clearly
    useful anywhere promises were involved.
