+++ lib/bin/backup.js
@@ -10,7 +10,7 @@ const { google } = require('googleapis');
 const Problem = require('../util/problem');
 const Option = require('../util/option');
 const { connect } = require('../model/database');
-const { getOrElse } = require('../util/http');
+const { getOrElse } = require('../util/promise');
 const { generateLocalCipherer } = require('../util/crypto');
 const pkg = require('../model/package');
