+++ lib/cli.js
@@ -19,7 +19,8 @@ const ensurePassword = (input, f) => {
 // handles query finalization and cli output.
 const { compose } = require('ramda');
 const { inspect } = require('util');
-const { serialize, finalize } = require('./util/http');
+const { finalize } = require('./http/endpoint');
+const { serialize } = require('./util/http');
 const writeTo = (output) => (x) => output.write(`${x}\n`);
 const writeToStderr = writeTo(process.stderr);
 const printError = (error) => {
@@ -45,7 +46,7 @@ cli.parse(cliArgs, cliCommands);
 
 ////////////////////////////////////////////////////////////////////////////////
 // execute commands.
-const { getOrNotFound } = require('./util/http');
+const { getOrNotFound } = require('./util/promise');
 cli.main((args, options) => {
   if (cli.command === 'user-create') {
     ensurePassword(options.password, (password) =>
