+++ lib/http/service.js
@@ -18,7 +18,7 @@ module.exports = (container) => {
 
   // pull session and version information and pass it on, or reject the request if
   // broken information is provided for either.
-  const { versionParser, sessionParser, fieldKeyParser, headerOptionsParser } = require('../util/http');
+  const { versionParser, sessionParser, fieldKeyParser, headerOptionsParser } = require('./middleware');
   service.use(versionParser);
   service.use(sessionParser(container));
   service.use(fieldKeyParser(container));
@@ -40,7 +40,7 @@ module.exports = (container) => {
   // POSTRESOURCE HANDLERS
 
   // apply output error handler to everything.
-  const { sendError } = require('../util/http');
+  const { sendError } = require('./endpoint');
   service.use((error, request, response, next) => {
     if (response.headersSent === true) {
       // In this case, we'll just let Express fail the request out.
