+++ lib/model/query/actors.js
@@ -1,5 +1,5 @@
 const Problem = require('../../util/problem');
-const { getOrReject } = require('../../util/http'); // TODO: clearly, this is useful beyond http.
+const { getOrReject } = require('../../util/promise');
 
 module.exports = {
   create: (actor) => ({ actees, simply }) =>
