+++ lib/resources/config.js
@@ -1,10 +1,11 @@
 const { google } = require('googleapis');
 const config = require('config');
 const { isBlank, printPairs } = require('../util/util');
-const { endpoint, getOrNotFound, getOrElse, success } = require('../util/http');
+const { endpoint } = require('../http/endpoint');
+const { success } = require('../util/http');
 const Problem = require('../util/problem');
 const Option = require('../util/option');
-const { ExplicitPromise, reject } = require('../util/promise');
+const { ExplicitPromise, reject, getOrNotFound, getOrElse } = require('../util/promise');
 const { generateKeypair } = require('../util/crypto');
 
 // TODO: not here.
