+++ lib/resources/field-keys.js
@@ -1,4 +1,6 @@
-const { endpoint, getOrNotFound, success } = require('../util/http');
+const { endpoint } = require('../http/endpoint');
+const { getOrNotFound } = require('../util/promise');
+const { success } = require('../util/http');
 
 // This resource pertains specifically to /global/ field keys, which are permitted
 // to submit data to any form in the system. Field keys which are scoped to a given
