+++ lib/resources/forms.js
@@ -1,5 +1,6 @@
 const path = require('path');
-const { endpoint, openRosaEndpoint, getOrNotFound } = require('../util/http');
+const { endpoint, openRosaEndpoint } = require('../http/endpoint');
+const { getOrNotFound } = require('../util/promise');
 const { formList } = require('../outbound/openrosa');
 
 module.exports = (service, { Actee, Form, Audit }) => {
