+++ lib/resources/sessions.js
@@ -1,7 +1,9 @@
 const Problem = require('../util/problem');
 const { isBlank } = require('../util/util');
 const { verifyPassword } = require('../util/crypto');
-const { endpoint, getOrReject, success } = require('../util/http');
+const { endpoint } = require('../http/endpoint');
+const { getOrReject } = require('../util/promise');
+const { success } = require('../util/http');
 
 module.exports = (service, { User, Session }) => {
