+++ lib/resources/submissions.js
@@ -1,9 +1,9 @@
 const { readFileSync } = require('fs');
 const { always } = require('ramda');
 const sanitize = require('sanitize-filename');
-const { openRosaEndpoint, endpoint, getOrNotFound, getOrReject, rejectIf } = require('../util/http');
+const { openRosaEndpoint, endpoint } = require('../http/endpoint');
 const { openRosaMessage } = require('../outbound/openrosa');
-const { resolve, reject } = require('../util/promise');
+const { getOrNotFound, getOrReject, rejectIf, resolve, reject } = require('../util/promise');
 const Option = require('../util/option');
 const Problem = require('../util/problem');
 const { streamJoinedCsvs } = require('../data/csv');
