+++ lib/resources/users.js
@@ -1,7 +1,8 @@
-const { endpoint, getOrNotFound, success, isTrue } = require('../util/http');
+const { endpoint } = require('../http/endpoint');
+const { success, isTrue } = require('../util/http');
 // const { verifyPassword } = require('../util/crypto');
 const Problem = require('../util/problem');
-const { resolve, reject } = require('../util/promise');
+const { resolve, reject, getOrNotFound } = require('../util/promise');
 
 module.exports = (service, { Actee, User, mail }) => {
