+++ lib/util/promise.js
@@ -1,5 +1,9 @@
 const { curry, compose, identity } = require('ramda');
+const Problem = require('./problem');
 
+////////////////////////////////////////////////////////////////////////////////
+// EXPLICIT PROMISE
+//
 // An ExplicitPromise is a promise whose computation is not triggered until a
 // point() method is called. Resolution is thereupon passed up the entire promise
 // tree and executed starting from the rootmost point.
@@ -53,8 +57,31 @@ class MappedExplicitPromise extends ExplicitPromise {
   }
 }
 
+
+////////////////////////////////////////////////////////////////////////////////
+// PROMISE HELPERS
+// just like Promise.reject and Promise.resolve but gives ExplicitPromises.
+
 const reject = Promise.reject.bind(Promise);
 const resolve = (x) => ExplicitPromise.of(Promise.resolve(x));
 
-module.exports = { isExplicitPromise, maybePoint, ExplicitPromise, MappedExplicitPromise, reject, resolve };
+
+////////////////////////////////////////////////////////////////////////////////
+// .THEN HELPERS
+// things that work well with .then; eg: .then(getOrNotFound)
+
+const getOrElse = (orElse) => (option) => option.orElse(orElse);
+
+// can't use option.orElse here, as constructing a reject is necessarily a rejection.
+const getOrReject = (rejection) => (option) => (option.isDefined() ? option.get() : reject(rejection));
+const getOrNotFound = getOrReject(Problem.user.notFound());
+
+const rejectIf = (predicate, problem) => (value) => ((predicate(value) === true) ? reject(problem(value)) : value);
+
+
+module.exports = {
+  isExplicitPromise, maybePoint, ExplicitPromise, MappedExplicitPromise,
+  reject, resolve,
+  getOrElse, getOrReject, getOrNotFound, rejectIf
+};
